+++ src/components/MainContent.vue
@@ -1,3 +1,5 @@
+<!-- We have a bug where the page loaded on a phone looks horrible, wider than the screen. Might be tailwind error here -->
+
 <template>
     <div class="text-dark font-poppins text-m max-w-280 mx-auto ">
         <slot></slot>
